Fixed a precision issue in the interpolator to get a more precise time percentage.

+++ Vixen.System/Interpolator/Interpolator.cs
@@ -8,7 +8,7 @@ using Vixen.Sys.Attribute;
 namespace Vixen.Interpolator {
 	public abstract class Interpolator<T> {
 		public bool Interpolate(TimeSpan timeOffset, TimeSpan timeSpan, T startValue, T endValue, out T value) {
-			float percent = (float)(timeOffset.TotalMilliseconds / timeSpan.TotalMilliseconds);
+			double percent = timeOffset.TotalMilliseconds / timeSpan.TotalMilliseconds;
 			return Interpolate(percent, startValue, endValue, out value);
 		}
